// src/api-client/models/color-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ColorResponse
 */
export interface ColorResponse {
    /**
     * 
     * @type {number}
     * @memberof ColorResponse
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof ColorResponse
     */
    'name'?: string;
}